refactor: move layouts foldre from components to src
