+++ src/layouts/RootLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
-import Home from '../../pages/Home/Home';
 import styles from './RootLayout.module.scss';
+import Home from '../pages/Home/Home';
 
 export default function RootLayout() {
   return (
