+++ src/pages/SideCardDetails/SideCardDetails.module.scss
@@ -2,6 +2,8 @@
   position: sticky;
   display: flex;
   flex-direction: column;
+  justify-content: center;
+  align-items: center;
   gap: 20px;
   top: 10px;
   padding: 50px 20px;
@@ -11,6 +13,7 @@
   background-color: #393939;
 
   .CardDescriptionWrapper {
+    width: 100%;
     p {
       color: #b1a9a9;
       span {
